userSection added

+++ src/app/api/announcements/[id]/route.js
@@ -1,3 +1,4 @@
+import { getUser } from '@/lib/getUser';
 import prisma from '@/lib/prismaDB'
 import { NextResponse } from 'next/server'
 
@@ -27,6 +28,9 @@ export const GET = async (_req, { params }) => {
 export const DELETE = async (_req, { params }) => {
     const { id } = params;
     try {
+        const sessiion = await getUser()
+        if (sessiion.user.role !== process.env.ADMIN_ROLE)
+            return NextResponse.json({ message: "Unuthorized", error }, { status: 400 })
         await prisma.announcements.delete({
             where: { id }
         })
@@ -39,6 +43,9 @@ export const DELETE = async (_req, { params }) => {
 export const PATCH = async (req, { params }) => {
     const { id } = params;
     try {
+        const sessiion = await getUser()
+        if (sessiion.user.role !== process.env.ADMIN_ROLE)
+            return NextResponse.json({ message: "Unuthorized", error }, { status: 400 })
         const body = await req.json()
         const { title, description, imageURL, tags, createdBy, telegram, status, views } = body;
         const updateAnnouncements = await prisma.announcements.update({
